+++ src/Transformer/Exception/ClassNotInstantiableException.php
@@ -23,7 +23,7 @@ class ClassNotInstantiableException extends NotMappableValueException
     public function __construct(string $class, Context $context = null)
     {
         parent::__construct(
-            message: sprintf('Trying to instantiate class "%s", but this class is not instantiable.', $class),
+            message: sprintf('Trying to instantiate "%s", but it is not instantiable. You might solve this problem by adding an "InheritanceMap" to the class, so the mapper will know which concrete class to instantiate.', $class),
             context: $context
         );
     }